requirements.txt
hydra
hydra-core

pytorch
torchvision
torchaudio
cudatoolkit 11.3

omegaconf
tqdm

ftfy
lpips

regex